driver changes and I did mitch's mom

+++ src/main/java/edu/thangiah/action/driver/DeleteAction.java
@@ -41,7 +41,7 @@ public class DeleteAction extends ManagementController
     	}
     	
     	try{
-    		if(fromDb != null) {
+    		if(fromDb != null && driverDao.findByMaintenanceOrder(maintenanceOrder) == null) {
         		driverDao.delete(fromDb);    			
     		}
     	}
